+++ extension/helper/config.js
@@ -1,7 +1,7 @@
 /**
  * @param {ExtensionConfig} config
  */
-function getBaseUrl(config) {
+function getBaseUrl (config) {
   const url = config.shopifyShopDomain || 'https://' + config.shopifyShopAlias + '.myshopify.com'
   return url.replace(/\/$/, '')
 }