+++ extension/models/Errors/ScriptNotWhitelistedError.js
@@ -4,7 +4,7 @@ class ScriptNotWhitelistedError extends Error {
   constructor (scriptName) {
     super(scriptName
       ? `The script '${scriptName}.js' is not whitelisted.`
-      : `The given script is not whitelisted.`
+      : 'The given script is not whitelisted.'
     )
 
     this.code = ESCRIPTNOTWHITELISTED