+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- backend support for Shopify Selling Plans / Subscriptions
+
+### Fixed
+- display of "pseudo" variant products (only one option and one value to choose from) in the cart
 
 ## [3.4.0] - 2025-10-23
 ### Added